refactor: UserDetails 관련 코드 수정 (#26)
