+++ src/main/java/com/sjiwon/studyholic/security/principal/UserAuthenticationDto.java
@@ -1,4 +1,4 @@
-package com.sjiwon.studyholic.domain.entity;
+package com.sjiwon.studyholic.security.principal;
 
 import com.sjiwon.studyholic.domain.entity.user.User;
 import lombok.AccessLevel;
@@ -9,22 +9,22 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-public class AbstractUserPrincipal {
+public class UserAuthenticationDto {
     private Long id;
     private String name;
     private String nickname;
     private String loginId;
-    private String password;
+    private String loginPassword;
     private String email;
     private String storageName;
     private String role;
 
-    public AbstractUserPrincipal(User user) {
+    public UserAuthenticationDto(User user) {
         this.id = user.getId();
         this.name = user.getName();
         this.nickname = user.getNickName();
         this.loginId = user.getLoginId();
-        this.password = user.getLoginPassword();
+        this.loginPassword = user.getLoginPassword();
         this.email = user.getEmail();
         this.storageName = user.getStorageName();
         this.role = user.getRole().getAuthority();
